Add cross-boundary HarmonicBond and HarmonicAngleForces
Add harmonic bond and angle interactions across the QM-MM boundary.
Harmonic bond and angle forces are defined in the OpenMM force field
file, but bonds within the same residue are automatically zeroed in
PyDFT-QMMM; this change rectifies that. This will also add bonds
across the boundary where they may not exist in the OpenMM topology,
e.g., in a surface where atoms are separate residues.