Update link atom plugin to work with new version of PyDFT-QMMM.
Make various changes to the link-atom plugin to be compatible with the overhauled version of PyDFT-QMMM. Notably, rework things to no longer rely on the Psi4Context, which no longer exists.

Note: The code runs, but we still need to do physical testing to ensure the link-atom plugin is working properly. Also, this may or may not work with the new PME implementation.